Revision 3 (#2)
Built off of the original Revision 1 (and an intermediary Rev 2) of the Alkalinity Titrator PCB.

* Expanded PCB to include all breakout boards and appropriate interfacing
* Reworked the pH probe circuit to increase resistance to noise and increase precision
* Changed PT1000 connections to 3.5mm jack interface
* Added BD65496MUV dc motor driver for magnetic stirrer
* Added LCD, keypad, and magnetic stirrer components

The PCB has been largely revised starting in Fall. Revision 1 started with a simple PCB based off the original Alkalinity Titrator prototype, created by Greg Birge. Revision 2 was an updated board that was fabricated, tested, and then reworked into the current Revision 3. Most features of the Rev 2 and Rev 3 are the same, though a series of mistakes were corrected.

The board measures 85x100mm and is made to mount directly on top of the Raspberry Pi 3b/4b. It requires a dual channel 5V and 12V power supply, with the MEAN WELL RD-65A selected as the current supply to test with.

Co-authored-by: Konrad McClure <Konrad.McClure@wallawalla.edu>
Co-authored-by: Noah-Griffith <noah.griffith@wallawalla.edu>